fix: 서비스 목적에 맞게 코드 재부여
